+++ plex_metadata/guid.py
@@ -1,104 +1,128 @@
-from plex_metadata.core.defaults import DEFAULT_GUID_MAP, DEFAULT_TV_AGENTS
-from plex_metadata.core.helpers import try_convert, compile_map, urlparse
+from plex_metadata.agents import Agents
+from plex_metadata.core.helpers import urlparse
 
 import logging
 
 log = logging.getLogger(__name__)
+unsupported_agents = {}
 
 
 class Guid(object):
-    map = compile_map(DEFAULT_GUID_MAP)
-
-    def __init__(self, agent, sid, extra=None):
-        self.agent = agent
-        self.sid = sid
+    def __init__(self, value=None, extra=None, matched=False, invalid=False, agent_id=None, original=None):
+        self.value = value
         self.extra = extra
+
+        self.matched = matched
+        self.invalid = invalid
+
+        self.agent_id = agent_id
+        self.original = original
+
+        # Identifier
+        self.service = None
+        self.id = None
 
         # Show
         self.season = None
         self.episode = None
 
+        # Optional
+        self.language = None
+
+    @property
+    def agent(self):
+        return self.service
+
+    @property
+    def sid(self):
+        # `sid` property always returns strings
+        return str(self.id)
+
+    @property
+    def valid(self):
+        return not self.invalid and self.matched
+
     @classmethod
-    def parse(cls, guid, map=True):
-        if not guid:
-            return None
+    def construct(cls, service, id, extra=None, matched=False, invalid=False):
+        result = cls(
+            id, extra,
+            matched=matched,
+            invalid=invalid
+        )
 
-        agent, uri = urlparse(guid)
-
-        if not agent or not uri or not uri.netloc:
-            return None
-
-        result = Guid(agent, uri.netloc, uri.query)
-
-        # Nothing more to parse, return now
-        if uri.path:
-            cls.parse_path(result, uri)
-
-        if map:
-            return cls.map_guid(result)
-
+        result.service = service
+        result.id = id
         return result
 
     @classmethod
-    def parse_path(cls, guid, uri):
-        # Parse path component for agent-specific data
-        path_fragments = uri.path.strip('/').split('/')
+    def parse(cls, guid, match=True, media=None, strict=False):
+        if not guid:
+            return cls(
+                invalid=True,
+                original=guid
+            )
 
-        if guid.agent in DEFAULT_TV_AGENTS:
-            if len(path_fragments) >= 1:
-                guid.season = try_convert(path_fragments[0], int)
+        # Parse Guid URI
+        agent_id, uri = urlparse(guid)
 
-            if len(path_fragments) >= 2:
-                guid.episode = try_convert(path_fragments[1], int)
-        else:
-            log.warn('Unable to completely parse guid (agent: %r)', guid.agent)
+        if not agent_id or not uri or not uri.netloc:
+            return cls(
+                invalid=True,
+                agent_id=agent_id,
+                original=guid
+            )
+
+        # Construct `Guid` object
+        result = cls(
+            uri.netloc,
+            extra=uri.query,
+
+            agent_id=agent_id,
+            original=guid
+        )
+
+        # Match guid with agent, fill with details
+        if match and cls.match(agent_id, result, uri, media):
+            result.matched = True
+            return result
+
+        if strict:
+            return result
+
+        # No agent matching enabled, automatically match guid parameters
+        log.warn('Unable to find agent mapping for %s://%s, result may be incorrect', agent_id, uri.netloc)
+
+        result.service = agent_id[agent_id.rfind('.') + 1:]
+        result.id = uri.netloc
+        return result
 
     @classmethod
-    def map_guid(cls, guid):
-        agent, sid_pattern, match = cls.find_map(guid)
+    def match(cls, agent_name, guid, uri, media=None):
+        # Retrieve `Agent` for provided `guid`
+        agent = Agents.get(agent_name)
 
-        guid.agent = agent
+        if agent is None:
+            if agent_name not in unsupported_agents:
+                # First occurrence of unsupported agent
+                log.warn('Unsupported metadata agent: %s' % agent_name)
 
-        # Match sid with regex
-        if sid_pattern:
-            if not match:
-                log.warn('Failed to match "%s" against sid_pattern for "%s" agent', guid.sid, guid.agent)
-                return None
+                # Mark unsupported agent as "seen"
+                unsupported_agents[agent_name] = True
+                return False
 
-            # Update with new sid
-            guid.sid = ''.join(match.groups())
+            # Duplicate occurrence of unsupported agent
+            log.warn('Unsupported metadata agent: %s' % agent_name, extra={
+                'duplicate': True
+            })
+            return False
 
-        return guid
-
-    @classmethod
-    def find_map(cls, guid):
-        # Strip leading key
-        agent = guid.agent[guid.agent.rfind('.') + 1:]
-
-        # Return mapped agent and sid_pattern (if present)
-        mappings = cls.map.get(agent, [])
-
-        if type(mappings) is not list:
-            mappings = [mappings]
-
-        for mapping in mappings:
-            map_agent, map_pattern = mapping
-
-            if map_pattern is None:
-                return map_agent, None, None
-
-            match = map_pattern.match(guid.sid)
-            if not match:
-                continue
-
-            return map_agent, map_pattern, match
-
-        return agent, None, None
+        # Fill `guid` with details from agent
+        return agent.fill(guid, uri, media)
 
     def __repr__(self):
         parameters = [
-            'agent: %r' % self.agent,
-            'sid: %r' % self.sid
+            'service: %r' % self.service,
+            'id: %r' % self.id
         ]
 
         if self.season is not None: