+++ CHANGES.rst
@@ -1,7 +1,43 @@
+0.8.0 (2016-09-06)
+------------------
+**Added**
+ - Guid
+    - Implemented new :code:`Guid` agent mapper
+    - Support for strict guid parsing
+    - :code:`Guid.parse()` will now always return a :code:`Guid` object, use the :code:`Guid.valid` property to check validity
+    - Properties
+       - :code:`agent_id`
+       - :code:`id`
+       - :code:`invalid`
+       - :code:`language`
+       - :code:`matched`
+       - :code:`original`
+       - :code:`service`
+       - :code:`valid`
+       - :code:`value`
+    - Agent Mappings:
+       - :code:`com.arendshome.plex.agents.personalmedia`
+       - :code:`com.plexapp.agents.AlloCine`
+       - :code:`com.plexapp.agents.anidb`
+       - :code:`com.plexapp.agents.cinepassion`
+       - :code:`com.plexapp.agents.filmaffinity`
+       - :code:`com.plexapp.agents.hama` *(anidb, imdb, tmdb, tvdb, tvdb2, tvdb3)*
+       - :code:`com.plexapp.agents.kinopoisk`
+       - :code:`com.plexapp.agents.kinopoiskru`
+       - :code:`com.plexapp.agents.kinopoiskrushow`
+       - :code:`com.plexapp.agents.youtube`
+       - :code:`net.devvsbugs.coding.plex.myanimelist`
+       - :code:`net.fribbtastic.coding.plex.myanimelist`
+
+**Changed**
+ - Metadata
+    - Ensure the "key" provided to the :code:`fetch()` method is an integer
+    - Removed some unneeded logger messages
+
 0.7.1 (2016-02-12)
 ------------------
 **Changed**
- - Reduced severity of `Item ... with type "..." has been ignored` message
+ - Reduced severity of :code:`Item ... with type "..." has been ignored` message
  - Updated travis configuration (python: dropped 3.2, added 3.4 and 3.5)
 
 **Fixed**
@@ -44,4 +80,4 @@
 
 0.6.0 (2014-10-30)
 ------------------
-- Initial release
+ - Initial release